Service: Avoid rejecting when service is unidled
When a service is back from an idle state, the corresponding Load
Balancer should have the reject field set to false, as the client
that triggered the unidling process is likely to retransmit an SYN
TCP packet, as per the retry mechanism.
During the 30 seconds time slot (called "Grace Period") after the
unidling trigger, incoming connection are not rejected by the Load
Balancers, giving the endpoints the time to come up.

Add a controller to mark unidled services with
`k8s.ovn.org/unidled-at` annotation, and use that to build load
balancer options.

Avoid setting NB_Global.Options["controller_event"] as we need to
not raising events when LoadBalancers has `event=false` and
`reject=false`.

Add `SetAnnotationOnService(...)` method to kube.Kube as helper
function.

Adjusting unit and e2e tests to the new logic.

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>