Ensure VF representor is not managed by OVS before renaming.
If the representor is managed by OVS when renaming, it will cause an inconsistency
inside OVS between the userspace and datapath, resulting in a bad situation where
the interface has to be renamed to the original value before the port can be
properly removed from the userspace and datapath.

This patch bypasses the issue by deleting the port inside OVS if it exists.

Signed-off-by: Alin-Gabriel Serdean <aserdean@ovn.org>