Remove MTU on service routes
Now that OVN is capable of sending ICMP needs frag for PMTU discovery,we
no longer need to proxy that capability in the host.

However, there seems to be a bug with this feature when trying to send
OVN a packet that is too large:
https://bugzilla.redhat.com/show_bug.cgi?id=2169839

This commit only changes to remove the MTU specified on the
169.254.169.1 route. This is the OVNMasquerade IP and is only used when
hairpinning traffic back to the node which will end up at a host network
pod. Therefore there is no reason to restrict its MTU as we are never
forwarding to an OVN networked pod in this case.

Fixes: #3419

Signed-off-by: Tim Rozet <trozet@redhat.com>