rm metallb repo in destroy_metallb() function
if that repo exists when re-running kind it will throw and error
during the install_metallb() step when it tries to clone it.
the repo is removed later in the install_metallb() step, but
if something breaks or exits before that happens it could be
leftover making the clone step on the next install fail. You
could check if the repo folder exists before hand, but then
you would not be using the latest version of the repo... or
would have to add steps to pull the latest which could get
more complicated if files are potentially modified. this
is simple enough.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>