netpol: Manage multiple deletion of a pod
In case of the gressPolicy `deletePeerPod` is invoked
multiple times with Pods that share the same IP, the
related Address_Set may ends in an inconsistent state.

Refactor to BaseNetworkController.findPodWithIPAddresses(...) to have
a single function to check for pod IP duplication.

Add a check in DefaultNetworkController.handlePeerPodSelectorDelete(...)
to be sure we are not deleting an IP address that is used by another pod

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>