Fix NAD controller Stop
NAD controller would wait for the workers to be terminated before
shutting down the worker queue. In turn the workers could be endlessly
waiting for the queue to be shut down. This results in the NAD
controller never actually stopping.

Improved also how the network controllers are stopped for which we
wait until the NAD controller workers have effectively terminated.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>