Depend only on ovnkube master LE metrics
Currently, we have two LE processes in main,
however, LE framework only allows setting
a metric provider once for these LE metrics,
and since CM is set first, its LE metrics
are used and NCM are ignored.

Lets depend on NCM LE metrics for now as
downstream, we depend on this for alerts.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>