Add Device Plugin Resource For Mgmt Port
A new argument ovnkube-node-mgmt-port-dp-resource-name is added
in this commit to pass the device plugin resource name that would
be used for the management port. The resource will override the
existing parameter ovnkube-node-mgmt-port-netdev, if provided.

A new DPResourceDeviceIdsMap is added in this commit to the
OvnKubeNodeConfig structure to track the resources and allocated
device ID that would be used for mgmt port and other future
features such as bypass port.

Signed-off-by: William Zhao <wizhao@redhat.com>