lgw: append init rules instead insert
The rules that are created on init should be appended instead
of inserted as they are "lowest priority".
More specifically, if they are inserted before EGRESS-SVC it breaks
because its rules are never evaluated - this may happen if the
init rules do not exist but the jump to EGRESS-SVC does (e.g when moving
from shared to local mode).

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>