unidling: Keep `event=true` during grace period
During the grace perdio (30s after a service is unidled) the load balancer
must not send TCP rejects, not raise any empty_lb_event and it can have
no backends. Because of [1], this configuration is not supported by OVN
and as a workaround we keep `event=true`. This can generate multiple
NeedPods events, but they will be discarded by the unidling controller.

Refs:
[1] https://bugzilla.redhat.com/show_bug.cgi?id=2177173

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>