Update DefaultDeny ACLs to have new dbIDs. Add NetpolNamespaceOwnerType for namespace-wide default deny acls.
Add acl_sync PrimaryID check to make sure in case more than 1 ACLs have
the same primaryID, only of them will be updated.
This also eliminates the need for "deleting a network policy that
failed half-way through creation succeeds" test, which is based on a
"found multiple results for provided predicate" condition.

Replace tests
"stale ACLs should be cleaned up or updated at startup via
syncNetworkPolicies" and
"ACLs with long names and run syncNetworkPolicies"
with the new "reconciles an existing networkPolicy updating stale
ACLs with long names", and a part that deletes stale ARP ACLs is now
updated and tested as a part of acl_sync.

Signed-off-by: Nadia Pinaeva <npinaeva@redhat.com>