CARRY: use "prefer local" for annotated services
Openshift hack: the router service is annotated with "traffic-policy.network.alpha.openshift.io/local-with-fallback" because, while it is defined with externalTrafficPolicy=Local, it expects the CNI to switch to externalTrafficPolicy=Cluster as soon as there are no local endpoints on any given node. SDN does this already, let's do it in OVNK as well to further reduce disruption time to openshift ingress during upgrades.

Co-authored-by: Riccardo Ravaioli <rravaiol@redhat.com>
Co-authored-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>