Enable forwarding on br-ex && br-ex1
Since we are adding a new disable-forwarding
flag for all OVNK interfaces, we want to
enable forwarding for our primary and secondary
bridge interfaces that we own explicitly.

NOTE: This commit has no relevance upstream since
we anyways don't disable forwarding on our interfaces.
It's more for downstream.

NOTE2: We are doing this for DPU mode as well, if that
is not needed we can remove doing that in the future,
I don't know enough about that mode to make that
determination here.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>