multi-net, kind: provide multus daemonset spec
OVN-K provides a 0.4.0 CNI configuration; the default multus daemonset
spec provides a 0.3.1 configuration.

Furthermore, there's a feature in which multus rejects handling any
requests for deletegates in versions supporting CNI CHECK ( >= 0.4.0)
when the multus configuration does **not** handle CNI CHECK ( < 0.4.0).

Thus, we need to maintain the multus CNI configuration, since we need to
configure it's CNI version to be compliant with our delegate
configuration.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>