Expose PfID and VfID in DPU host mode and read that in DPU mode
Currently, the vf rep needs to be specified manually on the DPU side,
but given a configuration on the host side, there is only one correct
configuration on the DPU side. Instead of relying on the user to provide
the right config on both sides, expose the config on the host side
through an annotation and read that annotation from the DPU side.

Signed-off-by: Balazs Nemeth <bnemeth@redhat.com>