multi-homing, policy, e2e: add test with namespaceSelectors
Adds tests (both for flat L2 and routed topologies) asserting only pods
from the desired namespaces can contact a certain workload.

For this test, we had to introduce a separate namespace, in which we
will instantiate the "allowed" workload. To simplify the test
setup/teardown, this happens in `BeforeEach` / `AfterEach` blocks, for
which we had to create a separate context.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>