Add EgressService "Network" E2E
Verifying that the "Network" field works properly:
Setting a "dummy" custom routing table on all of the nodes (this heavily relies on the environment to be a kind cluster):
We create a new routing table with 2 routes to the external container:
1) The one from the default routing table.
2) A blackhole with a higher priority
Then in the actual test we first verify that when the pods are using the custom routing table they can't reach the external container,
remove the blackhole route and verify that they can reach it now. This shows that they actually use a different routing table than the main one.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>