e2e cp tests: Strip the leading and trailing qoutes from stdout of a cmd.
The function findOvnKubeMasterNode() returns the name of the master
node with the leading and trailing quotes because of which the tests
which are suppose to delete the pods running on the master node doesn't
actually delete any pods.

This patch fixes the issue.  For single node zone deployments,
ovnkube-ncm will be running on each node and the lease by the
name 'ovn-kubernetes-master-global' doesn't exist. Hence
the function findOvnKubeMasterNode() also tries to get the lease by the
name 'ovn-kubernetes-master-ovn-control-plane' if getting the lease
by the name 'ovn-kubernetes-master-global' fails.

Signed-off-by: Numan Siddique <numans@ovn.org>