Enable github actions merge group
This will help resolve the issue of merging a PR then requiring another
PR to be rebased to ensure we test against new HEAD. PRs will be sent to
a merge queue and then executed against HEAD before they are merged.

Once this commit is merged we can go enable requiring merge queue in the
repo.

Signed-off-by: Tim Rozet <trozet@redhat.com>