Allow external gateway bridge without uplink port In local gateway mode
Users can create the external gateway bridge without attaching a
host physical interface as the uplink port in local gateway mode.

The gateway router will use 169.254.169.4 as the default gateway.

Add a new gateway config flag 'allow-no-uplink' to controll if this
setup is allowed or not. It's disabled by default.

With this setup, egressIP and egressGW can not work.

Signed-off-by: Peng Liu <pliu@redhat.com>