dbchecker: don't delete database file; just back it up
Deleting the database file also makes the server forget its cluster ID
and server ID. This causes the DB to make a new cluster ID and server ID
on restart, which actually causes split-brain, because the other servers
in the cluster think there's a new member at the same IP address, but the
ond member isn't removed. This either causes them to reject the new
member or to think the cluster has more members than it should.

When trying to reset the RAFT cluster state, all DBs on all masters
need to be removed at the same time, and all servers must be restarted
at the same time so that the don't keep the old cluster ID and server
IDs. The dbchecker cannot do that, because it only runs on a single
machine and can't coordinate between nodes.

Deleting the DB just makes the problem worse and we've see it actually
cause split-brain in a number of cases where the DBs could have
recovered.

https://issues.redhat.com/browse/OCPBUGS-11705

Signed-off-by: Dan Williams <dcbw@redhat.com>