Call SyncEndpoints from AddService
There is a race condition observed when the
endpointslice is added before the service.
We will not pass
https://github.com/ovn-org/ovn-kubernetes/blob/e7a92dbdab7d2f0c056ae2c6e549140c1809c7c0/go-controller/pkg/node/healthcheck_service.go#L131
this check and hence will not call `syncEndPointSlices`
since service doesn't exist yet. Later when we do
AddService, we were not calling syncEndpoints and so
we end up in a situation where the localEndpoints count
is not updated.

This PR fixes that by moving the logic to call syncEndpoints
to AddService. In the reverse case of endpointslice being added
after the service, its possible we call the syncEndpoints from
two places - both AddService and AddEndpointSlice, but better
to be functionally correct and avoid race before worrying
about performance...

(cherry picked from commit dea653ab12d058fcc011529a2d1792cb69fc0a28)
Conflicts because https://github.com/ovn-org/ovn-kubernetes/pull/3485
does not exist.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>