Re-structure ipallocator package
Move pkg/ipallocator to pkg/allocator/ip
Move pkg/ipallocator/allocator to pkg/allocator/bitmap

This makes sense since other components are using the allocator package
without it actually being associated to allocating IPs.

It also facilitates proper placement of a future subnet IP allocator.

Had to fix an issue in hack/test-go.sh where it was using a recursive
grep to find ginkgo references and pass ginkgo arguments. This grep
should not be recursive, otherwise it will pass ginkgo arguments if a
subpackage has ginkgo references but not the root package which was the
case for pkg/allocater after this change.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>