Add a subnet ip allocator
Move most of the logical switch manager functionality to an independant
subnet ip allocator so it can be used from different places without it
being associated with the concept of switches or nodes.

Logical switch manager retains the functionality of reserving the
gateway, management and hybrid overlay IPs for every allocated subnet.

Everything is functionally equivalent except when reserving the hybrid
overlay IP. An issue was fixed where it could be releasing an IP it had
not allocated in the first place.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>