Pre-allocate subnets in subnet ip allocator
Add the ability to pre-allocate subnets to the subnet ip allocator as it
is a common need: we use this for a variety of pruposes, including
reserving well known addresses in the default network or L3 secondary
networks, as well as reserving user requested address for L2 secondary
networks.

Note that this is just a pre-allocation. The subnet ip allocator does
not distinguish between pre-allocated IPs vs normally allocated IPs and
will return the same ErrAllocated error on re-allocation.

In the process, refactor GetIPFullMask to be based on net.IPMask and
avoided parsing in a few places.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>