Move subnet handling out of network cluster controller
Network cluster controller will be in charge of coordinating more than
just host subnets so move that functionality to the host subnet allocator
itself. That functionality fits in the host subnet allocator which
it was almost only a façade to the base subnet allocator.

Functionaly equivalent.

Fixes subnet count metric being used for all networks instead of just
the default network.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>