Add ip allocator to cluster manager
Adds an ip allocator to cluster manager. The allocator reconciles pods
triggered by events registered by the cluster manager network controller
which now listens to pod events when interconnect is enabled.

The ip allocator not only allocates IPs but fully handles the
PodAnnotation as well. It will derive from the allocated IPs or honor
requests for the mac address and gateways and set the annotation on the
Pod. This will trigger the CNI side to attach the pod interface while
master network controller configures the northbound database.

Only enabled for localnet secondary networks with interconnect for
now. On such a scenario, the master network controller is inhibited from
setting the PodAnnotation leaving it up for the cluster manager to do.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>