EgressService: sync node labels
Make sure the egress-service labels of the node match only the services allocated to it.
Up until now we only added and deleted labels according to our caches, but we didn't
handle the situation where a label that does not belong to the node appears on it
(e.g by a user adding it manually).

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>