fix issue running tests in docker
something has changed in the interaction between git and
golang 1.18.5 and how it works with docker. It's not a
problem in 1.18.4 and also not a problem running tests
with podman.

adding the flag to not buildvcs when invoking go will
resolve the issue. It's believed there are no consumers
of this VCS information in the UTs so no need to build
with them.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
        modified:   hack/test-go.sh

no changes added to commit (use "git add" and/or "git commit -a")
root@f4e9245e2268:/go/src/github.com/ovn-org/ovn-kubernetes/go-controller#

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>