Fixes usages of get pod IPs
We were treating a failure to get IPs on a pod (because it had no IPs at
the time) as an error in certain non-pod handlers. This would cause
entire syncs or adding handlers to fail, when it shouldn't because once
the handler/sync has been done the pod watcher will see the update event
when the pod finally gets its IP address.

Also, its possible to get a delete event for a pod that never had an IP
address but was scheduled.

Signed-off-by: Tim Rozet <trozet@redhat.com>
Co-authored-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>