kind: Change the first zone name from zone-1 to global.
For multinode per zone interconnect deployments change the
zone name from zone-1 to global.  Its good to have one
zone with the default zone name - global.  This unearths a
bug in ovnkube-controller. During the deployment, the
ovnkube-controller running on global zone (ovn-control-plane node)
when it starts adds the nodes - ovn-worker2 and ovn-worker3 to
its local zone name map and later when ovnkube-nodes running
on ovn-worker2 and ovn-worker3 nodes updates the zone annotation
of these nodes to "zone-2", ovnkube-controller of global zone
(in the function addUpdateRemoteNodeEvent()) fails to
    (1) cleanup the node resources for these nodes
    (2) create interconnect resources for these nodes
    (3) remove these nodes from the local zone name map

Because of (3), ovnkube-controller is in a continuous retry loop until its
restarted.

Note: This commit doesn't fix the issue. Just reproduces the issue in
CI.  The CI job - e2e (multi-node-zones, noHA, local, ipv4, SnatGW, 1br, ic-multi-node-zones, 3, 2)
fails because of this bug.

Signed-off-by: Numan Siddique <numans@ovn.org>