e2e multi-homing: fix network name changing throughout the test
Every time attachamentName() was being used, a new network name was
being generated. If used more than once throughout a single test, the
different parts of that test would use a different network name.
Specifically for localnet tests, this meant that the bridge mapping
would not be set with the same network name as the NAD network name.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>