Add support for zones
Most of this feature has no dependency on IC
or multi-zones or single zones or anything. It is
unfortunately one of those features that won't benefit
scale wise from using IC mode.

However the two parts we care about for which we need to
introduce the concept of zones are:

a) adding ports to port groups - we need to do this
only for local pods
b) status update support - comes in upcoming commits

The commit also takes care of passing the feature flag
into the IC deployment mode.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>