Add NCMWatchFactory
We have 3 modes that we can run in:
a) cluster-manager
b) network controller manager
c) node

When IC=true and multiple zones are used
cluster-manager and network-controller-manager are run
separately. CloudPrivateIPConfigType is used only
from cluster-manger in that case and there is no
need to initialize it from ncm.

However when these modes are run together, we use
the master watch factory and this creates confusion.
Here is what this PR proposes:

a) CMWatch Factory when CM is run independently
b) NCMWatch Factory when NCM is run independently
c) NodeWatch Factory when node is run independently
d) MasterWatch Factory when CM+NCM are run together
e) MasterWatch Factory when NCM+Node are run together
f) MasterWatch Factory when NCM+Node+CM are run together

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>