Account for draft PRs in test workflows
PRs will be set as draft when opened.

A draft PR will only run lint and unit test jobs.

A PR has to be flagged as ok-to-review to run e2e tests.

A specific job will fail for draft PRs to avoid a time window when a PR
could be merged without being tested when flagged as ok-to-review.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>