Run unit test on a separate workflow
This allows to easily run e2e tests independently of the unit tests and
with different trigger conditions.

We still need to upload the image artifact both from the unit tests and
the e2e workflows as the upload is per-workflow. In the case of the e2e
workflow, we take care to remove the image when done to save space.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>