Set mac binding age threshold in gateway routers
On scenarios where the cluster is shut down and started up after some
time later, the dynamic mac binding entries in the SBDB might be stale.
If any endpoint changed MAC address and sent a GARP, OVN has no chance
of processing it if it is shut down.

Set mac_binding_age_threshold to 300 seconds. Most common values used in
other implementations range from 60s to 1200s being 300s the most common
after a google search.

Unfortunately the mac binding age of an entry is not refreshed on any
network event, so entries will be removed from the mac binding table
after 300 seconds unconditionaly. There is an initiative to improve this
behavior in OVN [1]

[1] https://mail.openvswitch.org/pipermail/ovs-discuss/2023-May/052475.html

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>